requirements.txt
requests
pyyaml
boto3
docker
flask
flask-cors
flask-jwt-extended
psutil
